Create url_utils.py to prevent circular imports
By introducing `parse_bool()` function and unifying the environment
variable parsing, we would have created circular import between rhsm's
config.py and utils.py, if `in_container()` also used it.

As the file contains several utility functions that are focused on
handling URLs, they have been moved to separate file. One of these
functions is `get_env_proxy_info()`, which was the only function in
utils.py that was using `parse_url()`.

By moving it away the circular import was resolved.

List of changes:
- Added `parse_bool()` in `rhsm/config.py`.
- Added source file rhsm/url_utils.py
- Moved URL functions from rhsm/utils.py to rhsm/url_utils.py
- Added test file rhsm/unit/test_url_utils.py
- Moved URL tests to rhsm/unit/test_url_utils.py
